refactor : 공지 생성 검증 위치 변경
